Update unit tests against new collector layout; format the main package to run the new collector code.
