+++ collector/mesos_agent/mesos_agent.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 	"html/template"
 	"net"
+	"net/http"
 	"net/url"
 	"reflect"
 	"strconv"
@@ -26,6 +27,7 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
+	"github.com/dcos/dcos-metrics/collector"
 	"github.com/dcos/dcos-metrics/collector/http_client"
 	"github.com/dcos/dcos-metrics/producers"
 )
@@ -39,8 +41,115 @@ var (
 		"collector": "mesos-agent"})
 )
 
+/* MesosAgentCollector defines the collector type for Mesos
+agent. It is configured from main from config file options and
+pass a new instance of HTTP client and a channel for dropping metrics
+onto */
+type MesosAgentCollector struct {
+	Port            int           `yaml:"port"`
+	PollPeriod      time.Duration `yaml:"poll_period"`
+	RequestProtocol string        `yaml:"request_protocol"`
+
+	MetricsChan chan producers.MetricsMessage
+	NodeInfo    collector.NodeInfo
+	HTTPClient  *http.Client
+
+	agentState       agentState
+	containerMetrics []agentContainer
+}
+
+// agentContainer defines the structure of the response expected from Mesos
+// *for a single container* when polling the '/containers' endpoint in API v1.
+// Note that agentContainer is actually in a top-level array. For more info, see
+// https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/agent/agent.proto#L161-L168
+type agentContainer struct {
+	FrameworkID     string                 `json:"framework_id"`
+	ExecutorID      string                 `json:"executor_id"`
+	ExecutorName    string                 `json:"executor_name"`
+	Source          string                 `json:"source"`
+	ContainerID     string                 `json:"container_id"`
+	ContainerStatus map[string]interface{} `json:"container_status"`
+	Statistics      *resourceStatistics    `json:"statistics"`
+}
+
+// agentState defines the structure of the response expected from Mesos
+// *for all cluster state* when polling the /state endpoint.
+// Specifically, this struct exists for the following purposes:
+//
+//   * collect labels from individual containers (executors) since labels are
+//     NOT available via the /containers endpoint in v1.
+//   * map framework IDs to a human-readable name
+//
+// For more information, see the upstream protobuf in Mesos v1:
+//
+//   * Framework info: https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/mesos.proto#L207-L307
+//   * Executor info:  https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/mesos.proto#L474-L522
+//
+type agentState struct {
+	ID         string          `json:"id"`
+	Hostname   string          `json:"hostname"`
+	Frameworks []frameworkInfo `json:"frameworks"`
+}
+
+type frameworkInfo struct {
+	ID        string         `json:"id"`
+	Name      string         `json:"name"`
+	Principal string         `json:"principal"`
+	Role      string         `json:"role"`
+	Executors []executorInfo `json:"executors"`
+}
+
+type executorInfo struct {
+	ID        string           `json:"id"`
+	Name      string           `json:"name"`
+	Container string           `json:"container"`
+	Labels    []executorLabels `json:"labels,omitempty"` // labels are optional
+}
+
+type executorLabels struct {
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}
+
+// resourceStatistics defines the structure of the response expected from Mesos
+// when referring to container and/or executor metrics. In Mesos, the
+// ResourceStatistics message is very large; it defines many fields that are
+// dependent on a feature being enabled in Mesos, and not all of those features
+// are enabled in DC/OS.
+//
+// Therefore, we redefine the resourceStatistics struct here with only the fields
+// dcos-metrics currently cares about, which should be stable for Mesos API v1.
+//
+// For a complete reference, see:
+// https://github.com/apache/mesos/blob/1.0.1/include/mesos/v1/mesos.proto#L921-L1022
+type resourceStatistics struct {
+	// CPU usage info
+	CpusUserTimeSecs      float64 `json:"cpus_user_time_secs,omitempty"`
+	CpusSystemTimeSecs    float64 `json:"cpus_system_time_secs,omitempty"`
+	CpusLimit             float64 `json:"cpus_limit,omitempty"`
+	CpusThrottledTimeSecs float64 `json:"cpus_throttled_time_secs,omitempty"`
+
+	// Memory info
+	MemTotalBytes uint64 `json:"mem_total_bytes,omitempty"`
+	MemLimitBytes uint64 `json:"mem_limit_bytes,omitempty"`
+
+	// Disk info
+	DiskLimitBytes uint64 `json:"disk_limit_bytes,omitempty"`
+	DiskUsedBytes  uint64 `json:"disk_used_bytes,omitempty"`
+
+	// Network info
+	NetRxPackets uint64 `json:"net_rx_packets,omitempty"`
+	NetRxBytes   uint64 `json:"net_rx_bytes,omitempty"`
+	NetRxErrors  uint64 `json:"net_rx_errors,omitempty"`
+	NetRxDropped uint64 `json:"net_rx_dropped,omitempty"`
+	NetTxPackets uint64 `json:"net_tx_packets,omitempty"`
+	NetTxBytes   uint64 `json:"net_tx_bytes,omitempty"`
+	NetTxErrors  uint64 `json:"net_tx_errors,omitempty"`
+	NetTxDropped uint64 `json:"net_tx_dropped,omitempty"`
+}
+
 func (h *MesosAgentCollector) RunPoller() {
-	ticker := time.NewTicker(h.PollPeriod)
+	ticker := time.NewTicker(h.PollPeriod * time.Second)
 
 	// Poll once immediately
 	h.pollMesosAgent()
@@ -65,7 +174,7 @@ func (h *MesosAgentCollector) pollMesosAgent() {
 	// always fetch/emit agent state first: downstream will use it for tagging metrics
 	maColLog.Debugf("Fetching state from DC/OS host %s", host)
 	if err := h.getAgentState(); err != nil {
-		maColLog.Errorf("Failed to get agent state from %s. Error: %s", host, h.Port, err)
+		maColLog.Errorf("Failed to get agent state from %s. Error: %s", host, err)
 	}
 
 	maColLog.Debugf("Fetching container metrics from DC/OS host %s", host)
@@ -79,7 +188,7 @@ func (h *MesosAgentCollector) pollMesosAgent() {
 func (h *MesosAgentCollector) getContainerMetrics() error {
 	h.containerMetrics = []agentContainer{}
 	u := url.URL{
-		Scheme: "http",
+		Scheme: h.RequestProtocol,
 		Host:   net.JoinHostPort(h.NodeInfo.IPAddress, strconv.Itoa(h.Port)),
 		Path:   "/containers",
 	}
@@ -96,7 +205,7 @@ func (h *MesosAgentCollector) getAgentState() error {
 	h.agentState = agentState{}
 
 	u := url.URL{
-		Scheme: "http",
+		Scheme: h.RequestProtocol,
 		Host:   net.JoinHostPort(h.NodeInfo.IPAddress, strconv.Itoa(h.Port)),
 		Path:   "/state",
 	}
